Rename project to DODEC
