+++ audit/dodec/src/PerformAggregation.go
