+++ audit/dodec/src/aggregations/GetCollectionCount.go
@@ -2,11 +2,11 @@ package aggregations
 
 import (
 	"context"
+	"dodec/types"
 	"fmt"
 	"go.mongodb.org/mongo-driver/v2/bson"
 	"go.mongodb.org/mongo-driver/v2/mongo"
 	"log"
-	"pull-audit-data/types"
 )
 
 // GetCollectionCount returns a `simpleMap` data structure as defined in the PerformAggregation function. Each key is
