+++ audit/dodec/src/aggregations/GetDocsIdsWithRecentActivity.go
@@ -2,10 +2,10 @@ package aggregations
 
 import (
 	"context"
+	"dodec/types"
 	"fmt"
 	"go.mongodb.org/mongo-driver/v2/bson"
 	"go.mongodb.org/mongo-driver/v2/mongo"
-	"pull-audit-data/types"
 	"time"
 )
