+++ audit/dodec/src/aggregations/GetLanguageCounts.go
@@ -2,10 +2,10 @@ package aggregations
 
 import (
 	"context"
+	"dodec/types"
 	"go.mongodb.org/mongo-driver/v2/bson"
 	"go.mongodb.org/mongo-driver/v2/mongo"
 	"log"
-	"pull-audit-data/types"
 )
 
 // GetLanguageCounts returns a `simpleMap` data structure as defined in the PerformAggregation function. Each key is the
