+++ audit/dodec/src/aggregations/GetMinMedianMaxCodeLength.go
@@ -2,10 +2,10 @@ package aggregations
 
 import (
 	"context"
+	"dodec/types"
 	"go.mongodb.org/mongo-driver/v2/bson"
 	"go.mongodb.org/mongo-driver/v2/mongo"
 	"log"
-	"pull-audit-data/types"
 )
 
 // GetCodeLengths returns a `lengthCountMap` data structure as defined in the PerformAggregation function.
