+++ audit/dodec/src/aggregations/GetSubProductCategoryCounts.go
